Fixed Malloc Error
Evidently you shouldn't allocate sizeof(foo*) instead of sizeof(foo).
Runs to completion but need to verify logic. Strong suspicion that the
logic for extractresults needs to be upset to offset the buffers we are
writing to as we currently probably just grab the first batch of
requests every iteration